Use Decimal type for event_id

+++ model.rb
@@ -6,7 +6,7 @@ class ToriBird
   include DataMapper::Resource
   property :id, Serial
   property :etag, String, :length => 256, :required => false
-  property :event_id, Integer, :required => false
+  property :event_id, Decimal, :required => false
   property :access_token, String, :length => 256, :required => false
 end
